Refactor old and new MCTS and Node
