+++ project2/mcts_playground.py
@@ -3,7 +3,7 @@ import logging
 from project2.config import logging_config
 from project2.games.catch_game_state_manager import CatchGameStateManager
 from project2.logging_config import configure_logging
-from project2.mcts_new import MCTSNew
+from project2.mcts import MCTS
 from project2.models.perfect_model import PerfectModel
 
 try:
@@ -18,12 +18,13 @@ gs_initial_state = gs_manager.generate_initial_state()
 gs_available_actions = gs_manager.get_legal_actions(gs_initial_state)
 
 model = PerfectModel(gs_manager)
-mcts = MCTSNew(model, gs_available_actions, simulations=500)
+mcts = MCTS(model, gs_available_actions, simulations=800, discount=1)
 
 initial_state = model.represent_state(gs_initial_state)
 available_actions = gs_manager.get_legal_actions(initial_state)
 state = initial_state
 done = False
+total_reward = 0
 
 print("Begin")
 print(state)
@@ -36,3 +37,6 @@ while not done:
     available_actions = gs_manager.get_legal_actions(state)
     print(f"New state. Reward {reward} from action {action}\n", state)
     print()
+    total_reward += reward
+
+print(f"Total reward: {total_reward}")
